[Refactoring] Split the implementation of compiling MPS and BNNS
