+++ services/ml/BUILD.gn
@@ -66,8 +66,14 @@ source_set("lib") {
     sources += [
       "compilation_impl_mac.h",
       "compilation_impl_mac.mm",
+      "compilation_impl_mac_bnns.h",
+      "compilation_impl_mac_bnns.mm",
+      "compilation_impl_mac_mps.h",
+      "compilation_impl_mac_mps.mm",
       "execution_impl_mac.h",
       "execution_impl_mac.mm",
+      "ml_utils_mac.h",
+      "ml_utils_mac.mm",
       "model_impl_mac.h",
       "model_impl_mac.mm",
       "mpscnn_context.h",
