+++ services/ml/compilation_impl_mac_bnns.h
@@ -1,0 +1,41 @@
+// Copyright 2017 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef SERVICES_ML_COMPILATION_IMPL_MAC_BNNS_H_
+#define SERVICES_ML_COMPILATION_IMPL_MAC_BNNS_H_
+
+#include <map>
+#include <memory>
+#include <vector>
+
+#include "services/ml/common.h"
+#include "services/ml/ml_utils_mac.h"
+
+namespace ml {
+
+bool CompileConv2DBNNS(OperationMac&,
+                       std::map<uint32_t, ValueInfo>& values,
+                       std::unique_ptr<int8_t[]>& memory,
+                       std::vector<OperandMac>& operands);
+
+bool CompileAverageOrMaxPool2DBNNS(OperationMac&,
+                                   std::map<uint32_t, ValueInfo>& values,
+                                   std::unique_ptr<int8_t[]>& memory,
+                                   std::vector<OperandMac>& operands);
+
+bool CompileSoftmaxBNNS(OperationMac&,
+                        std::map<uint32_t, ValueInfo>& values,
+                        std::unique_ptr<int8_t[]>& memory,
+                        std::vector<OperandMac>& operands);
+
+bool CompileReshapeBNNS(OperationMac&);
+
+bool CompileConcatenationBNNS(OperationMac& operation,
+                              std::map<uint32_t, ValueInfo>& values,
+                              std::unique_ptr<int8_t[]>& memory,
+                              bool is_concatenation_first);
+
+}  // namespace ml
+
+#endif  // SERVICES_ML_COMPILATION_IMPL_MAC_BNNS_H_
